Add standard T1 3/4 LED package
The existing part, led-5mm.stl, looks pretty, but doesn't actually
match the physical characteristics of a standard LED. This part
doesn't look pretty, but is as exact a match of a CAD drawing of
a standard T1 3/4 LED as I can make it.

I tried to export STL, but FreeCAD didn't export the file.